renamed constants.py into types.py
